Revert "Feature/navigator"
