+++ package.json
@@ -35,7 +35,6 @@
     "react-native-randombytes": "2.2.0",
     "react-native-tcp": "^2.0.4",
     "react-native-udp": "^1.2.9",
-    "react-navigation": "^1.0.0-beta.11",
     "readable-stream": "^1.0.33",
     "secp256k1": "3.3.0",
     "stream-browserify": "^1.0.0",
