+++ src/components/App.js
@@ -5,7 +5,15 @@ import {
   Text,
   View
 } from 'react-native';
-import WalletService from '../services/WalletService';
+
+import crypto from 'crypto';
+import bip39 from 'bip39';
+import buffer from 'buffer';
+import util from 'ethereumjs-util';
+// import lightwallet from 'eth-lightwallet'; // doesn't work =(
+import hdkey from 'ethereumjs-wallet/hdkey';
+import EthereumTx from 'ethereumjs-tx';
+
 
 const styles = StyleSheet.create({
   container: {
@@ -26,12 +34,62 @@ const styles = StyleSheet.create({
   },
 });
 
+/*
+  // Naranja usefull code ?)
+  self.pushContractTransaction = function(_keys, transaction) {
+    $log.info("[backend - pushContractTransaction] Starting", "transaction[" + angular.toJson(transaction) + ']');
+    return self.signHash(transaction.info.to_sign, _keys.privkey)
+    .then(function (signature) {
+      $log.info("[backend - pushContractTransaction] Info", "Calling apiCall(/push_tx)");
+      transaction.signature = signature.signature;
+      return self.apiCall(_keys, ENVIRONMENT.apiurl('/push_tx'), angular.toJson(transaction));
+    });
+  }
+  self.callContractFunction = function(_keys, token_id, fnc, payload) {
+    $log.info("[backend - callContractFunction] Starting", "function[" + fnc + "]" , "payload[" + angular.toJson(payload) + ']');
+    return self.apiCall(_keys, ENVIRONMENT.apiurl('/token/' + token_id + '/' + fnc), payload)
+    .then(function(transaction) {
+      return self.pushContractTransaction(_keys, transaction);
+    });
+  }
+*/
+
+
+const generateWallet = () => {
+  console.log(hdkey);
+  const randomBytes = crypto.randomBytes(16);
+  const mnemonic = bip39.entropyToMnemonic(randomBytes.toString('hex'));
+  const seed = bip39.mnemonicToSeed(mnemonic);
+  const ethKey = hdkey.fromMasterSeed(seed);
+  const wallet = ethKey.getWallet();
+  return {
+    mnemonic: mnemonic,
+    privateKey: wallet.getPrivateKey(),
+    publicKey: wallet.getPublicKeyString().substr(2)
+  }
+};
+
 App = () => {
-  const wallet = WalletService.generateNewWallet();
-  console.log("wallet info:", wallet); 
-  const toAddress = "0x00c376412f3a8063fc6bceb1d874730ea88eb531";
-  const amount = 11;
-  WalletService.send({privateKey: wallet.privateKey, fromAddress: wallet.address}, toAddress, amount);
+  const wallet = generateWallet();
+  console.log("mnemonic", wallet.mnemonic); 
+
+  // dummy tx test:
+  const txParams = {
+    nonce: '0x00',
+    gasPrice: '0x09184e72a000', 
+    gasLimit: '0x2710',
+    to: '0x0000000000000000000000000000000000000000', 
+    value: '0x00', 
+    data: '0x7f7465737432000000000000000000000000000000000000000000000000000000600057',
+    // EIP 155 chainId - mainnet: 1, ropsten: 3
+    chainId: 3
+  };
+
+  const tx = new EthereumTx(txParams);
+  tx.sign(wallet.privateKey);
+  const serializedTx = tx.serialize();
+  console.log("serializedTx", serializedTx);
+
   return (
     <View style={styles.container}>
       <Text style={styles.welcome}>
