--- src/services/API/API.js
@@ -1,8 +1,0 @@
-
-import contractAPI from './contract.js';
-
-const API = {};
-
-API.contract = contractAPI;
-
-export default API;
